requirements.txt
Django==2.2.6
django-nested-admin==3.2.4
pip==19.0.3
python-monkey-business==1.0.0
pytz==2019.3
selenium==3.141.0
sqlparse==0.3.0
urllib3==1.25.6
sentry-sdk==0.14.1
psycopg2>=2.7,<3.0
django-solo==1.1.3
django-chroniker==1.0.17
mysqlclient==1.4.6
celery==4.4.7
django-celery-results==1.2.1
django-celery-beat==2.0.0
redis==3.5.3  # https://github.com/antirez/redis